Replace all the tests and make them all pass.

+++ src/apps/intel_mp/selftest.sh
@@ -2,17 +2,28 @@
 cd $(dirname $0)
 [ -z $SNABB_PCI_INTEL1G0 ] && exit $TEST_SKIPPED
 [ -z $SNABB_PCI_INTEL1G1 ] && exit $TEST_SKIPPED
-TESTS=$(find . -executable | grep -Pe 'test\d+\.' | sort)
+[ -z $SNABB_PCI_INTEL0 ] && exit $TEST_SKIPPED
+[ -z $SNABB_PCI_INTEL1 ] && exit $TEST_SKIPPED
+FILTER=${1:-.*}
+TESTS=$(find . -executable | grep -e 'test[0-9]' -e 'test_' | grep -e "$FILTER" | sort)
 ESTATUS=0
 for i in $TESTS; do
+   pkill -f snabb
+   sleep 1
+   rm -f /var/run/snabb/intel_mp*
+   rm -f results.*
    $i
    if test $? -eq 0; then
-      echo -n "PASSED"
+      echo "PASSED: $i"
    else
-      echo -n "FAILED"
+      for res in `ls results.*`; do
+	      echo $res;
+	      cat $res
+	      echo
+      done
+      echo "FAILED: $i"
       ESTATUS=-1
    fi
-   echo " :$i"
+   sleep 1
 done
-rm -f results.*
 exit $ESTATUS
